Fix directories
